requirements.txt

Django==4.2.7
django-environ==0.11.2
djangorestframework==3.14.0
drf-spectacular==0.26.5
flake8==6.1.0
mysqlclient==2.2.0
python-decouple==3.8
django-debug-toolbar==4.2.0
django-seed==0.3.1